renamed user.js to userServices.js
